requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.csv==1.0.7
boto==2.49.0
boto3==1.18.15
botocore==1.21.15
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
dealer==2.1.0
decorator==5.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.6
django-cors-headers==3.7.0
django-extensions==3.1.3
django-import-export==2.5.0
django-prettyjson==0.4.1
django-rest-registration==0.6.3
django-rest-swagger==2.2.0
django-ses==2.2.0
django-storages==1.11.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docutils==0.17.1
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
itypes==1.2.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
MarkupPy==1.14
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.7
parso==0.8.2
pickleshare==0.7.5
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
Pygments==2.9.0
PyJWT==2.1.0
python-dateutil==2.8.2
python-http-client==3.3.2
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
s3transfer==0.5.0
selenium==3.141.0
sendgrid==6.7.1
sentry-sdk==1.3.1
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
standardjson==0.3.1
starkbank-ecdsa==1.1.1
tablib==3.0.0
traitlets==5.0.5
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
wcwidth==0.2.5
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
